requirements.txt
bcrypt==4.0.1
blinker==1.6.2
cachelib==0.10.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
cryptography==41.0.3
dnspython==2.3.0
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Session==0.5.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
oauthlib==3.2.2
pycparser==2.21
pymongo==4.4.0
pyOpenSSL==23.2.0
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
urllib3==2.0.4
Werkzeug==2.3.6
